src/requirements.txt
kafka_python==2.0.2
loguru==0.6.0
protobuf<4
py4j==0.10.9.5
pyspark==3.3.1
redis==4.5.4
streamlit==1.20.0
google-cloud-bigquery==3.9.0
google-cloud-storage==2.8.0